view role and add role added

+++ Tracker/index.js
@@ -10,7 +10,7 @@ const connection = mysql.createConnection({
     database: "employees_db"
 });
 
-const viewEmployees = function (callback) {
+const viewEmployees = function () {
     let query = `select employee.id, employee.last_name, employee.first_name, role.title, department.name as department, role.salary
      from employee
      left join role on employee.role_id = role.id
@@ -21,7 +21,7 @@ const viewEmployees = function (callback) {
 
         console.table(res);
         
-        callback();
+        menu();
     });
 };
 
@@ -29,22 +29,28 @@ const addEmployee = function () {
     inquirer.prompt([questions[1], questions[2]]).then(res => {
         connection.query(`insert into employee (first_name, last_name) values (?, ?)`, [res.firstName, res.lastName], (err,result) => {
             if (err) throw err;
-            viewEmployees(menu);
+            viewEmployees();
         })
     });
 };
 
 const removeEmployee = function(employee) {
-    connection.query(`delete from employee where id=?`,[employee.id], (err, res) => {
-        if (err) throw err;
-        
-        viewEmployees(menu);
-    });
-}
+    if(employee)
+        connection.query(`delete from employee where id=?`,[employee.id], (err, res) => {
+            if (err) throw err;
+            
+            viewEmployees();
+        });
+};
 
 const selectEmployee = function (callback) {
     connection.query("select id, first_name, last_name from employee;", (err,res) => {
         if (err) throw err;
+
+        if(res.length==0) {
+            callback(false);
+            return;
+        }
 
         questions[3].choices=[];
         res.forEach(emp => {
@@ -57,7 +63,55 @@ const selectEmployee = function (callback) {
             callback({id: empID, first_name: name[1], last_name: name[0]});
         });
     });
-}
+};
+
+const viewRoles = function() {
+    connection.query("select * from role", (err,res) => {
+        if (err) throw err;
+
+        if (res.length == 0)
+            console.log('There are no roles yet');
+        console.table(res);
+        menu();
+    })
+};
+
+const addRole = function(department) {
+    inquirer.prompt([questions[4], questions[5]]).then(res => {
+        let query = 'insert into role (title, salary, department_id) values (?,?,null);';
+
+        if (department)
+            query=query.replace('null',`'${department.id}'`);
+            
+        connection.query(query, [res.title,res.salary], (err, res) => {
+            if (err) throw err;
+
+            viewRoles();
+        });
+    });
+
+};
+
+const selectDepartment = function (callback) {
+    connection.query("select * from department;", (err,res) => {
+        if (err) throw err;
+
+        if(res.length==0) {
+            callback(false);
+            return;
+        }
+
+        questions[6].choices=[];
+        res.forEach(dept => {
+            questions[6].choices.push(`${dept.id}. ${dept.name}`);
+        });
+        
+        inquirer.prompt(questions[6]).then(res => {
+            let dept=res.department.split('. ');
+            callback({id: dept[0], name: dept[1]});
+        });
+    });
+};
 
 const options = ['View All Employees', 'View All Employees by Dept', 'View All Employees by Manager', 'Add Employee', 'Remove Employee', 'Update Employee Role', 'Update Employee Manager', 'View All Roles', 'Add Role', 'Remove Role', 'View All Departments', 'Add Department', 'Remove Department', 'View Department Budget', 'Exit Program'];
 const questions = [{
@@ -80,6 +134,21 @@ const questions = [{
         name: 'employee',
         message: `Which employee?`,
         choices: []
+    }, {
+    // 4: addRole - title
+        name: 'title',
+        message: `What is the title? `
+    }, {
+    // 5: addRole - salary
+        type: 'number',
+        name: 'salary',
+        message: `What is the salary? `
+    }, {
+    // 6: selectDepartment - department
+        type: 'list',
+        name: 'department',
+        message: 'Which department?',
+        choices: []
     }
     ];
 
@@ -88,7 +157,7 @@ const menu = function () {
     inquirer.prompt(questions[0]).then(res => {
       switch(res.action) {
         case 'View All Employees':
-            viewEmployees(menu);
+            viewEmployees();
             break;
         case 'View All Employees by Dept':
             departmentSelection();
@@ -112,7 +181,7 @@ const menu = function () {
             viewRoles();
             break;
         case 'Add Role':
-            addRole();
+            selectDepartment(addRole);
             break;
         case 'Remove Role':
             removeRole();
